Implement analog read and color changing logic
The ADC is configured to read an analog value from RA4 using Vdd as a
voltage reference. That value is then converted to a number between 120
and 360 to be used as the hue value in HSV to RGB conversion.

Converting from HSV normally uses floating points, but to save
programming memory it is implemented as fixed point with two decimal
places.